FIX: static casts to double added

+++ Code/Numerics/itkMultivariateLegendrePolynomial.h
@@ -177,13 +177,15 @@ public:
         if (index[1] != m_PrevY)
           {
             // normalized y [-1, 1]
-            double norm_y = m_NormFactor[1] * index[1] - 1 ;
+            double norm_y = static_cast<double>(
+                              m_NormFactor[1] * index[1] - 1 );
             CalculateXCoef(norm_y, m_CoefficientArray) ;
             m_PrevY = index[1] ;
           }
         
         // normalized x [-1, 1]
-        double norm_x = m_NormFactor[0] * index[0] - 1 ;
+        double norm_x = static_cast<double>(
+                              m_NormFactor[0] * index[0] - 1 );
         
         return LegendreSum(norm_x, m_Degree, m_CachedXCoef) ;
       }
@@ -192,7 +194,8 @@ public:
         if (m_PrevZ != index[2])
           {
             // normalized z [-1, 1]  
-            double norm_z = m_NormFactor[2] * index[2] - 1 ;
+            double norm_z = static_cast<double>(
+                              m_NormFactor[2] * index[2] - 1 );
             CalculateYCoef(norm_z, m_CoefficientArray) ;
             m_PrevZ = index[2] ;
           }
@@ -200,7 +203,8 @@ public:
         if (m_PrevY != index[1])
           {
             // normalized y [-1, 1]
-            double norm_y = m_NormFactor[1] * index[1] - 1 ; 
+            double norm_y = static_cast<double>(
+                              m_NormFactor[1] * index[1] - 1 ); 
             CalculateXCoef(norm_y, m_CachedYCoef) ;
             m_PrevY = index[1] ;
           }
